Remove use of `.includes` to better support IE
This replaces the one use of `Array.prototype.includes` with
`Array.prototype.indexOf`, which can be used to achieve the same
thing: `x.includes(y) === x.indexOf(y) > -1`. This allows the code
to work in IE11 and earlier, which do not support `.includes`

I've also standardized the `browsers` targets across the `.babelrc`
and `roolup.config.js` files to make them consistent.