Actually use `markLoaded`'s params
`markLoaded` was being called with the same params regardless of
load success or error, hiding information about such errors. This
commit causes us to keep track of links loaded successfully or
with an error separately, allowing us to inform the user of the
links that caused the error. We may also in the future be able to
do things such as re-try attempts.