Add new method `onBeforeGetContent`
This method allows the user to change the contents of the page
before it is retrieved for printing. This occurs before the
existing `onBeforePrint` which is called after the page content has
been retrieved and right before the print window is opened.

Long term, it likely makes sense to change `onBeforePrint` to have
this functionality, and then remove this function, since there is
little use for it with this function existing.